requests 2.34 bugfix
